+++ Sources/MediaXDMEventGenerator.swift
@@ -22,9 +22,9 @@ class  MediaXDMEventGenerator {
     private let refEvent: Event
     private var lastReportedQoe: XDMQoeDataDetails?
     private var isTracking: Bool = false
-    private var refTS: Int64
+    private var refTS: Int
     private var currentPlaybackState: MediaContext.MediaPlaybackState?
-    private var currentPlaybackStateStartRefTS: Int64
+    private var currentPlaybackStateStartRefTS: Int
     private let allowedAdPingIntervalRangeInSeconds = 1...10
     private let allowedMainPingIntervalRangeInSeconds = 10...50
 
@@ -37,7 +37,7 @@ class  MediaXDMEventGenerator {
     #endif
 
     /// Initializes the Media XDM Event Generator
-    public required init(context: MediaContext, eventProcessor: MediaEventProcessing, config: [String: Any], refEvent: Event, refTS: Int64) {
+    public required init(context: MediaContext, eventProcessor: MediaEventProcessing, config: [String: Any], refEvent: Event, refTS: Int) {
         self.mediaContext = context
         self.mediaEventProcessor = eventProcessor
         self.trackerConfig = config
@@ -206,7 +206,7 @@ class  MediaXDMEventGenerator {
         addGenericDataAndProcess(eventType: XDMMediaEventType.statesUpdate, mediaCollection: mediaCollection)
     }
 
-    func setRefTS(ts: Int64) {
+    func setRefTS(ts: Int) {
         refTS = ts
     }
 
@@ -240,7 +240,7 @@ class  MediaXDMEventGenerator {
         // For bitrate change events and error events, use the qoe data in the current event being generated. For other events check MediaContext QoE object for latest QoE data updates.
         mediaCollection.qoeDataDetails = getQoEForCurrentEvent(qoe: mediaCollection.qoeDataDetails)
         // Add playhead details
-        mediaCollection.playhead = Int64(mediaContext.playhead)
+        mediaCollection.playhead = mediaContext.playhead
 
         // Convert the refTS from milliseconds to seconds
         let timestampAsDate = Date(timeIntervalSince1970: Double(refTS / 1000))
@@ -305,20 +305,20 @@ class  MediaXDMEventGenerator {
     /// Gets the custom reporting interval set in the tracker configuration. Valid custom main ping interval range is (10 seconds - 50 seconds) and valid ad ping interval is (1 second - 10 seconds)
     /// - Parameter isAdStart: A Boolean  when true denotes reporting interval is needed for Ad content or denotes Main content when false.
     /// - Return: the custom interval in `MILLISECONDS` if found in tracker configuration. Returns the default `MediaConstants.PingInterval.REALTIME_TRACKING` if the custom values are invalid or not found
-    private func getReportingIntervalFromTrackerConfig(isAdStart: Bool = false) -> Int64 {
+    private func getReportingIntervalFromTrackerConfig(isAdStart: Bool = false) -> Int {
         if isAdStart {
             guard let customAdPingInterval = trackerConfig[MediaConstants.TrackerConfig.AD_PING_INTERVAL] as? Int, allowedAdPingIntervalRangeInSeconds.contains(customAdPingInterval) else {
                 return MediaConstants.PingInterval.REALTIME_TRACKING_MS
             }
 
-            return Int64(customAdPingInterval) * 1000 // convert to Milliseconds
+            return customAdPingInterval * 1000 // convert to Milliseconds
 
         } else {
             guard let customMainPingInterval = trackerConfig[MediaConstants.TrackerConfig.MAIN_PING_INTERVAL] as? Int, allowedMainPingIntervalRangeInSeconds.contains(customMainPingInterval) else {
                 return MediaConstants.PingInterval.REALTIME_TRACKING_MS
             }
 
-            return Int64(customMainPingInterval) * 1000 // convert to Milliseconds
+            return customMainPingInterval * 1000 // convert to Milliseconds
         }
     }
 }
